ChangeSpeedBase in lib-builtin-effects
